Move path tests to tests/types/test_path.py
